requirements.txt
flask==3.1.0
gunicorn==23.0.0
requests==2.32.3
duckduckgo-search==7.5.3
google-analytics-data==0.18.16
google-auth==2.38.0
